add delete confirmation

+++ application/view/detail.php
@@ -20,7 +20,7 @@
 		<p>
 			asked by <?php echo $thread->email; ?> at <?php echo $thread->date; ?> | 
 			<a href="?controller=thread&action=formUpdate&query=<?php echo $thread->id; ?>">edit</a> | 
-			<a class="delete" href="delete" href="?controller=thread&action=delete&query=<?php echo $thread->id; ?>">delete</a>
+			<a class="delete" href="delete" href="?controller=thread&action=delete&query=<?php echo $thread->id; ?>" onclick="return confirmToDelete()">delete</a>
 		</p>
 	</div>	
 </div>
